some more styling changes mostly just fixed up bugs from the refactors
