+++ GUI/CreatePDF/create.js
@@ -1,92 +1,94 @@
-// Create PDF page JavaScript
-document.addEventListener('DOMContentLoaded', function() {
-    const frontImagesGrid = document.getElementById('frontImagesGrid');
-    const clearFrontBtn = document.getElementById('clearFrontBtn');
-    const createPdfBtn = document.getElementById('createPdfBtn');
-    const onlyFrontsCheckbox = document.getElementById('onlyFrontsCheckbox');
+const { ipcRenderer } = require('electron');
+const { FRONT_DIR } = require('../shared/constants');
+
+function loadImages() {
+    ipcRenderer.invoke('get-front-images').then(files => {
+        const grid = document.getElementById('frontImagesGrid');
+        grid.innerHTML = '';
+        files.forEach(f => {
+            const div = document.createElement('div');
+            div.style.textAlign = 'center';
+            div.style.wordBreak = 'break-all';
+            const img = document.createElement('img');
+            img.src = `${FRONT_DIR}/${f}`;
+            img.alt = f;
+            img.style.width = '100px';
+            img.style.height = '140px';
+            img.style.objectFit = 'cover';
+            img.style.borderRadius = '6px';
+            img.style.boxShadow = '0 2px 8px rgba(0,0,0,0.12)';
+            div.appendChild(img);
+            const label = document.createElement('div');
+            label.textContent = f;
+            label.style.fontSize = '0.8em';
+            label.style.marginTop = '0.5em';
+            div.appendChild(label);
+            grid.appendChild(div);
+        });
+    });
+}
+
+window.addEventListener('DOMContentLoaded', () => {
     const skip4Checkbox = document.getElementById('skip4Checkbox');
-    const pdfArgs = document.getElementById('pdfArgs');
 
-    // Load front images
-    function loadFrontImages() {
-        // Using Node.js fs module (Electron only)
-        const fs = window.require ? window.require('fs') : require('fs');
-        const path = window.require ? window.require('path') : require('path');
+    skip4Checkbox.addEventListener('change', function() {
+        let args = pdfArgsInput.value.trim();
+        const flag = '--skip 4';
+        if (this.checked) {
+            if (!args.includes(flag)) {
+                args = args.length ? args + ' ' + flag : flag;
+            }
+        } else {
+            // Remove the flag if present
+            args = args.replace(/\s*--skip 4\b/, '');
+        }
+        pdfArgsInput.value = args.trim();
+    });
+    const pdfArgsInput = document.getElementById('pdfArgs');
+    const onlyFrontsCheckbox = document.getElementById('onlyFrontsCheckbox');
+
+    onlyFrontsCheckbox.addEventListener('change', function() {
+        let args = pdfArgsInput.value.trim();
+        const flag = '--only_fronts';
+        if (this.checked) {
+            if (!args.includes(flag)) {
+                args = args.length ? args + ' ' + flag : flag;
+            }
+        } else {
+            // Remove the flag if present
+            args = args.replace(/\s*--only_fronts\b/, '');
+        }
+        pdfArgsInput.value = args.trim();
+    });
+    loadImages();
+    // Listen for event-driven updates from main process
+    ipcRenderer.on('front-images-changed', () => {
+        loadImages();
+    });
+    document.getElementById('createPdfBtn').onclick = async function() {
+        const args = document.getElementById('pdfArgs').value;
+        const spinner = document.getElementById('spinner');
+        spinner.classList.remove('hidden');
+        spinner.classList.add('flex');
         
-        const frontDir = path.resolve(__dirname, '../game/front');
-        if (!fs.existsSync(frontDir)) {
-            frontImagesGrid.innerHTML = '<p class="text-content-light/50 text-center p-4">No images found in front folder</p>';
-            return;
+        try {
+            const result = await ipcRenderer.invoke('run-create-pdf', args);
+            spinner.classList.remove('flex');
+            spinner.classList.add('hidden');
+            location.href = '../Utilities/pdf_viewer.html';
+        } catch (err) {
+            spinner.classList.remove('flex');
+            spinner.classList.add('hidden');
+            alert('Error creating PDF:\n' + err);
         }
-
-        const files = fs.readdirSync(frontDir).filter(file => 
-            file.toLowerCase().endsWith('.png') || 
-            file.toLowerCase().endsWith('.jpg') || 
-            file.toLowerCase().endsWith('.jpeg')
-        );
-
-        if (files.length === 0) {
-            frontImagesGrid.innerHTML = '<p class="text-content-light/50 text-center p-4">No images found in front folder</p>';
-            return;
-        }
-
-        frontImagesGrid.innerHTML = files.map(file => {
-            const filePath = `file://${path.join(frontDir, file)}`;
-            return `
-                <div class="card-grid-item">
-                    <img src="${filePath}" alt="${file}" loading="lazy" />
-                </div>
-            `;
-        }).join('');
     }
-
-    // Clear front images
-    clearFrontBtn.addEventListener('click', function() {
-        if (confirm('Are you sure you want to clear all front images?')) {
-            const fs = window.require ? window.require('fs') : require('fs');
-            const path = window.require ? window.require('path') : require('path');
-            const frontDir = path.resolve(__dirname, '../game/front');
-            
-            if (fs.existsSync(frontDir)) {
-                const files = fs.readdirSync(frontDir);
-                for (const file of files) {
-                    if (file !== 'EMPTY.md') {
-                        fs.unlinkSync(path.join(frontDir, file));
-                    }
-                }
-            }
-            loadFrontImages();
+    document.getElementById('clearFrontBtn').onclick = async function() {
+        try {
+            const result = await ipcRenderer.invoke('clear-front-images');
+            alert(result);
+            loadImages();
+        } catch (err) {
+            alert('Error clearing images:\n' + err);
         }
-    });
-
-    // Create PDF
-    createPdfBtn.addEventListener('click', function() {
-        const args = [];
-        if (onlyFrontsCheckbox.checked) args.push('--only_fronts');
-        if (skip4Checkbox.checked) args.push('--skip', '4');
-        if (pdfArgs.value.trim()) args.push(...pdfArgs.value.trim().split(' '));
-        
-        // Execute create_pdf.py with the arguments
-        const { spawn } = window.require ? window.require('child_process') : require('child_process');
-        const pythonProcess = spawn('python', ['create_pdf.py', ...args]);
-
-        pythonProcess.stdout.on('data', (data) => {
-            console.log(`stdout: ${data}`);
-        });
-
-        pythonProcess.stderr.on('data', (data) => {
-            console.error(`stderr: ${data}`);
-        });
-
-        pythonProcess.on('close', (code) => {
-            if (code === 0) {
-                alert('PDF created successfully!');
-            } else {
-                alert('Error creating PDF. Check the console for details.');
-            }
-        });
-    });
-
-    // Initial load
-    loadFrontImages();
+    }
 });
