+++ GUI/Home/home.html
@@ -6,7 +6,6 @@
     <title>Silhouette Card Maker - Home</title>
     <link rel="stylesheet" href="../styles/main.css">
     <script type="text/javascript" src="../shared/navbar.js"></script>
-    <script type="text/javascript" src="home.js"></script>
 </head>
 <body class="min-h-screen bg-surface text-content">
     <nav-bar current-step="1"></nav-bar>
