+++ GUI/MagicTheGathering/decklist.js
@@ -1,3 +1,4 @@
+const { FRONT_DIR } = require('../shared/constants');
 let cardData = [];
 let parsedCards = []; // Array of { name, setCode, setNumber, qty }
 
@@ -31,9 +32,8 @@ window.onload = function () {
     document.getElementById('exportImagesFSBtn').addEventListener('click', async function () {
         const fs = window.require ? window.require('fs') : require('fs');
         const path = window.require ? window.require('path') : require('path');
-        const frontDir = path.resolve(__dirname, '../game/front');
-        if (!fs.existsSync(frontDir)) {
-            fs.mkdirSync(frontDir, { recursive: true });
+        if (!fs.existsSync(FRONT_DIR)) {
+            fs.mkdirSync(FRONT_DIR, { recursive: true });
         }
         let imgCount = 1;
         for (let i = 0; i < cardData.length; i++) {
@@ -47,7 +47,7 @@ window.onload = function () {
                     let faceName = data.card_faces[f].name.replace(/[^a-zA-Z0-9]/g, ' ');
                     faceName = faceName.split(' ').map((w, idx) => idx === 0 ? w.toLowerCase() : w.charAt(0).toUpperCase() + w.slice(1).toLowerCase()).join('');
                     faceName = faceName.replace(/\s+/g, '');
-                    let filename = path.join(frontDir, `${imgCount}${faceName}1.png`);
+                    let filename = path.join(FRONT_DIR, `${imgCount}${faceName}1.png`);
                     const response = await fetch(imgUrl);
                     const buffer = Buffer.from(await response.arrayBuffer());
                     fs.writeFileSync(filename, buffer);
@@ -59,14 +59,14 @@ window.onload = function () {
                 let cardName = data.name.replace(/[^a-zA-Z0-9]/g, ' ');
                 cardName = cardName.split(' ').map((w, idx) => idx === 0 ? w.toLowerCase() : w.charAt(0).toUpperCase() + w.slice(1).toLowerCase()).join('');
                 cardName = cardName.replace(/\s+/g, '');
-                let filename = path.join(frontDir, `${imgCount}${cardName}1.png`);
+                let filename = path.join(FRONT_DIR, `${imgCount}${cardName}1.png`);
                 const response = await fetch(imgUrl);
                 const buffer = Buffer.from(await response.arrayBuffer());
                 fs.writeFileSync(filename, buffer);
                 imgCount++;
             }
         }
-        alert('Images exported to ../game/front');
+        alert('Images exported to ../../game/front');
     });
     // Export front images as ZIP button logic
     document.getElementById('exportImagesZipBtn').addEventListener('click', async function () {
