+++ GUI/Utilities/pdf_viewer.html
@@ -3,37 +3,26 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>PDF Viewer</title>
-    <style>
-        html, body { height: 100%; margin: 0; padding: 0; }
-        body { width: 100vw; height: 100vh; overflow: hidden; }
-        #backBtn {
-            position: absolute;
-            top: 20px;
-            left: 20px;
-            z-index: 10;
-            padding: 1em 2em;
-            font-size: 1.2em;
-            border-radius: 8px;
-            border: none;
-            background: #4b6cb7;
-            color: white;
-            cursor: pointer;
-            transition: background 0.2s;
-        }
-        #backBtn:hover { background: #182848; }
-        #pdfFrame {
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100vw;
-            height: 100vh;
-            border: none;
-        }
-    </style>
+    <title>Silhouette Card Maker - PDF Preview</title>
+    <link rel="stylesheet" href="../styles/main.css">
 </head>
-<body>
-    <button id="backBtn" onclick="location.href='home.html'">Back</button>
-    <iframe id="pdfFrame" src="../game/output/game.pdf"></iframe>
+<body class="min-h-screen bg-surface text-content">
+    <!-- Navigation Bar -->
+    <div class="fixed top-0 left-0 right-0 p-4 z-50 bg-surface/90 backdrop-blur-sm">
+        <div class="container mx-auto px-4">
+            <button 
+                onclick="location.href='../Home/home.html'" 
+                class="paper-button inline-flex items-center gap-2 !px-4 !py-2"
+            >
+                Back to Home
+            </button>
+        </div>
+    </div>
+
+    <!-- PDF Viewer -->
+    <iframe 
+        src="../../game/output/game.pdf" 
+        class="w-full h-screen border-none bg-surface"
+    ></iframe>
 </body>
 </html>
