+++ GUI/styles/main.css
@@ -1348,6 +1348,10 @@ html {
   height: 1rem;
 }
 
+.h-44 {
+  height: 11rem;
+}
+
 .h-5 {
   height: 1.25rem;
 }
@@ -1362,6 +1366,10 @@ html {
 
 .h-px {
   height: 1px;
+}
+
+.h-screen {
+  height: 100vh;
 }
 
 .max-h-\[400px\] {
@@ -1382,6 +1390,10 @@ html {
 
 .w-16 {
   width: 4rem;
+}
+
+.w-32 {
+  width: 8rem;
 }
 
 .w-4 {
@@ -1522,6 +1534,69 @@ html {
 
 .animate-bounce {
   animation: bounce 1s infinite;
+}
+
+@keyframes card-slide {
+  0% {
+    opacity: 0;
+    transform: translateY(-100%) translateX(-50%) rotate(-15deg);
+  }
+
+  25%, 35% {
+    opacity: 1;
+    transform: translateY(0) translateX(0) rotate(0deg);
+  }
+
+  60%, 100% {
+    opacity: 0;
+    transform: translateY(100%) translateX(50%) rotate(15deg);
+  }
+}
+
+.animate-card-slide-1 {
+  animation: card-slide 2s ease-in-out infinite;
+}
+
+@keyframes card-slide {
+  0% {
+    opacity: 0;
+    transform: translateY(-100%) translateX(-50%) rotate(-15deg);
+  }
+
+  25%, 35% {
+    opacity: 1;
+    transform: translateY(0) translateX(0) rotate(0deg);
+  }
+
+  60%, 100% {
+    opacity: 0;
+    transform: translateY(100%) translateX(50%) rotate(15deg);
+  }
+}
+
+.animate-card-slide-2 {
+  animation: card-slide 2s ease-in-out infinite 0.66s;
+}
+
+@keyframes card-slide {
+  0% {
+    opacity: 0;
+    transform: translateY(-100%) translateX(-50%) rotate(-15deg);
+  }
+
+  25%, 35% {
+    opacity: 1;
+    transform: translateY(0) translateX(0) rotate(0deg);
+  }
+
+  60%, 100% {
+    opacity: 0;
+    transform: translateY(100%) translateX(50%) rotate(15deg);
+  }
+}
+
+.animate-card-slide-3 {
+  animation: card-slide 2s ease-in-out infinite 1.33s;
 }
 
 .cursor-pointer {
@@ -2018,6 +2093,10 @@ html {
   gap: 1.5rem;
 }
 
+.gap-8 {
+  gap: 2rem;
+}
+
 .space-x-2 > :not([hidden]) ~ :not([hidden]) {
   --tw-space-x-reverse: 0;
   margin-right: calc(0.5rem * var(--tw-space-x-reverse));
@@ -2507,9 +2586,18 @@ html {
   border-style: none;
 }
 
+.border-content {
+  --tw-border-opacity: 1;
+  border-color: rgb(241 245 249 / var(--tw-border-opacity, 1));
+}
+
 .border-content-light {
   --tw-border-opacity: 1;
   border-color: rgb(203 213 225 / var(--tw-border-opacity, 1));
+}
+
+.border-content\/10 {
+  border-color: rgb(241 245 249 / 0.1);
 }
 
 .border-gray-200 {
@@ -2592,13 +2680,17 @@ html {
   background-color: rgb(22 24 29 / var(--tw-bg-opacity, 1));
 }
 
-.bg-surface-dark\/90 {
-  background-color: rgb(22 24 29 / 0.9);
+.bg-surface-dark\/95 {
+  background-color: rgb(22 24 29 / 0.95);
 }
 
 .bg-surface-light {
   --tw-bg-opacity: 1;
   background-color: rgb(42 48 60 / var(--tw-bg-opacity, 1));
+}
+
+.bg-surface\/90 {
+  background-color: rgb(30 35 47 / 0.9);
 }
 
 .decoration-slice {
@@ -2725,6 +2817,16 @@ html {
 
 .p-8 {
   padding: 2rem;
+}
+
+.\!px-4 {
+  padding-left: 1rem !important;
+  padding-right: 1rem !important;
+}
+
+.\!py-2 {
+  padding-top: 0.5rem !important;
+  padding-bottom: 0.5rem !important;
 }
 
 .px-0 {
@@ -3029,10 +3131,6 @@ html {
 .text-content-light {
   --tw-text-opacity: 1;
   color: rgb(203 213 225 / var(--tw-text-opacity, 1));
-}
-
-.text-content-light\/50 {
-  color: rgb(203 213 225 / 0.5);
 }
 
 .text-gray-500 {
@@ -3546,14 +3644,6 @@ html {
 
 .preserve-3d {
   transform-style: preserve-3d;
-}
-
-.\[animation-delay\:-\.3s\] {
-  animation-delay: -.3s;
-}
-
-.\[animation-delay\:-\.5s\] {
-  animation-delay: -.5s;
 }
 
 @media (min-width: 640px) {
